ENH: Allowing the variance to change with time
Providing a way to input variance that changes with time, such as
when an observatory drops out and we have a gap in data. The
SVD needs to be calculated anytime the variance is changed, but
that doesn't happen often, so don't just broadcast the SVD calls
through numpy which would calculate it at every timestep.