ENH: Add a new "variance" mode to the fit() function
Currently, we are keeping the components based on the ratio to the
maximum contributing variance. The new method calculates how much
variance each component contributes to the solution and selects the
number of components to keep (1-epsilon)% of the total variance.

This adds a new "mode" keyword argument that defaults to the previous
value, so is backwards compatible and simply adds a new method
of pruning values to the function.